requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

torch
transformers
Pillow
fastapi
uvicorn
timm
einops
python-multipart